Soc for PBFT
