+++ src/Components/Pages/Dashboard/Graphs/PBFT/data.js
@@ -1,0 +1,187 @@
+// ! POINTS DATA FOR REFERENCE FOR PLOTTING IS BELOW DO NOT DELETE IT!!!
+// const points = {
+//   request: {
+//     color: `${colors[0]}`,
+//     start: [
+//       {
+//         flag: true,
+//         points: { x: 0, y: 0 },
+//       },
+//     ],
+//     end: [
+//       {
+//         flag: true,
+//         points: { x: 200, y: 220 },
+//       },
+//     ],
+//   },
+//   prePrepare: {
+//     color: `${colors[1]}`,
+//     start: [
+//       {
+//         flag: true,
+//         points: { x: 200, y: 220 },
+//       },
+//     ],
+//     end: [
+//       {
+//         flag: true,
+//         points: { x: 400, y: 440 },
+//       },
+//       {
+//         flag: true,
+//         points: { x: 400, y: 660 },
+//       },
+//       {
+//         flag: true,
+//         points: { x: 400, y: 880 },
+//       },
+//     ],
+//   },
+//   prepare: {
+//     color: `${colors[2]}`,
+//     start: [
+//       { x: 400, y: 440 },
+//       { x: 400, y: 660 },
+//       { x: 400, y: 880 },
+//     ],
+//     end: [
+//       [
+//         {
+//           flag: true,
+//           points: { x: 600, y: 220 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 660 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 880 },
+//         },
+//       ],
+//       [
+//         {
+//           flag: true,
+//           points: { x: 600, y: 220 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 440 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 880 },
+//         },
+//       ],
+//       [
+//         {
+//           flag: true,
+//           points: { x: 600, y: 220 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 440 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 600, y: 660 },
+//         },
+//       ],
+//     ],
+//   },
+//   commit: {
+//     color: `${colors[3]}`,
+//     start: [
+//       { x: 600, y: 220 },
+//       { x: 600, y: 440 },
+//       { x: 600, y: 660 },
+//       { x: 600, y: 880 },
+//     ],
+//     end: [
+//       [
+//         {
+//           flag: true,
+//           points: { x: 800, y: 440 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 660 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 880 },
+//         },
+//       ],
+//       [
+//         {
+//           flag: true,
+//           points: { x: 800, y: 220 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 660 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 880 },
+//         },
+//       ],
+//       [
+//         {
+//           flag: true,
+//           points: { x: 800, y: 440 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 220 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 880 },
+//         },
+//       ],
+//       [
+//         {
+//           flag: true,
+//           points: { x: 800, y: 440 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 660 },
+//         },
+//         {
+//           flag: true,
+//           points: { x: 800, y: 220 },
+//         },
+//       ],
+//     ],
+//   },
+//   reply: {
+//     color: `${colors[4]}`,
+//     start: [
+//       {
+//         flag: true,
+//         points: { x: 800, y: 220 },
+//       },
+//       {
+//         flag: true,
+//         points: { x: 800, y: 440 },
+//       },
+//       {
+//         flag: true,
+//         points: { x: 800, y: 660 },
+//       },
+//       {
+//         flag: true,
+//         points: { x: 800, y: 880 },
+//       },
+//     ],
+//     end: [
+//       {
+//         flag: true,
+//         points: { x: 1000, y: 0 },
+//       },
+//     ],
+//   },
+// };
